refactor: move addSlot & deleteSlot to separate file
